Rename dijkstra files to shortest_path Closes #85
