+++ tests/graphs/mod.rs
@@ -1,3 +1,3 @@
 mod connected_components_tests;
-mod dijkstra_tests;
+mod shortest_path_tests;
 mod travelling_salesman_tests;
